fix: Put custom.scss at the end
The declarations in custom.scss are quite specific (not exactly
utilities, but close), and it is a good idea to put them after the
components. For example, a div with the banner and fullbleed classes
were not using the fullbleed because both classes were using width and
margin, and banner was declared after fullbleed. It was not the expected
the result, and with this change, we will make things more intuitive.